requirements.txt
google_trans_new==1.1.9
gtts==2.2.1
newspaper3k
textblob==0.15.3
streamlit==0.74.0
wikipedia==1.4.0